Added mainly database schema changes focused on relationships

+++ internal/models/game.go
@@ -7,7 +7,7 @@ import (
 type Game struct {
 	bun.BaseModel `bun:"table:games,alias:g"`
 
-	Number   uint8   `bun:",pk"`
+	Number   uint16  `bun:",pk"`
 	SeasonId int     `bun:",pk,notnull"`
 	Season   *Season `bun:"rel:belongs-to,join:season_id=id"`
 }
