+++ internal/models/player-log.go
@@ -6,7 +6,7 @@ type PlayerLog struct {
 	bun.BaseModel `bun:"table:players_logs,alias:pl"`
 
 	PlayerId      int     `bun:",pk,notnull"`
-	GameNumber    uint8   `bun:",pk,notnull"`
+	GameNumber    uint16  `bun:",pk,notnull"`
 	SeasonId      int     `bun:",pk,notnull"`
 	Player        *Player `bun:"rel:belongs-to,join:player_id=id"`
 	Game          *Game   `bun:"rel:belongs-to,join:game_number=number,join:season_id=season_id"`
